requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pydantic
python-dotenv

